+++ .env.example
@@ -1,3 +1,5 @@
+MIGRATIONS_PATH=migrations
+
 # mysql & postgres settings
 DB_DRIVER=postgres
 DB_HOST=localhost
