+++ .env.testing
@@ -1,3 +1,5 @@
+MIGRATIONS_PATH=migrations
+
 DB_DRIVER=sqlite3
 DB_FILE=test.db
 DB_CACHE=shared
