+++ examples/sql/20210315171956-create-users-table-down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS users;
