+++ examples/sql/20210315171956-create-users-table-up.sql
@@ -1,0 +1,7 @@
+CREATE TABLE IF NOT EXISTS users
+(
+    id bigserial not null primary key,
+    name varchar (255) not null,
+    email  varchar (255) not null unique,
+    created_at timestamp DEFAULT CURRENT_TIMESTAMP
+)
