+++ examples/sql/20210315174632-add-username-column-to-users-table-down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE users DROP COLUMN IF EXISTS username;
