+++ examples/sql/20210315174632-add-username-column-to-users-table-up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE users ADD COLUMN username varchar(255) not null unique;
