+++ pkg/entities/miigration.go
@@ -1,0 +1,10 @@
+package entities
+
+import "time"
+
+type MigrationEntity struct {
+	Id        uint      `db:"id"`
+	Migration string    `db:"migration"`
+	Version   uint      `db:"version"`
+	CreatedAt time.Time `db:"created_at"`
+}
